requirements.txt
aiohttp==3.5.4
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.4.4
chardet==3.0.4
discord==1.0.1
discord.py==1.2.5
idna==2.8
idna-ssl==1.1.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.6.1
pycodestyle==2.5.0
pylint==2.4.4
python-dotenv==0.10.3
six==1.13.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
websockets==6.0
wrapt==1.11.2
yarl==1.3.0
